model: data model for places to show in the neighborhood
Data model is hard coded for showing places in the neighborhood. The model is JSON formated and have a title for the archeological place, a kind to find-out what it is this place, the position to coordinate the excact position in maps and a short for shorthand name for this place.

Some additions in the places could be done.