feat: list of locations component and info box component
Two childs of the main app component created to render on screen the features of the list of places that marked on the map and the info box of the places when a place is clicked on mark or clicked on the elements of the list. Also, there is a search form to get results in every key input. The results is shown at the list items and at the markers on the map, too. Reseting the info box opened, by clicking somewhere empty area on the map.